+++ ExpenseTracker/Views/AdminSendNotification/CreateNotification.cshtml
@@ -1,0 +1,10 @@
+﻿@{
+    //ViewData["Title"] = "AdminSendNotification";
+    Layout = "_LayoutNoSidebar";
+
+}
+
+
+
+
+}
